+++ cmd/sebak/cmd/run.go
@@ -339,7 +339,7 @@ func runNode() error {
 	var g run.Group
 	{
 		nr, err := sebak.NewNodeRunner(flagNetworkID, localNode, policy, nt, isaac, st)
-		conf := &sebak.NodeRunnerConfiguration{
+		conf := &sebak.IsaacConfiguration{
 			TimeoutINIT:       timeoutINIT,
 			TimeoutSIGN:       timeoutSIGN,
 			TimeoutACCEPT:     timeoutACCEPT,
