+++ lib/isaac_configuration_test.go
@@ -1,5 +1,5 @@
 /*
-	In this file, there are unittests for checking NodeRunnerConfiguration struct.
+	In this file, there are unittests for checking IsaacConfiguration struct.
 */
 package sebak
 
@@ -12,7 +12,7 @@ import (
 
 //	TestConfigurationDefault tests the default timeout values.
 func TestConfigurationDefault(t *testing.T) {
-	n := NewNodeRunnerConfiguration()
+	n := NewIsaacConfiguration()
 	require.Equal(t, n.TimeoutINIT, 2*time.Second)
 	require.Equal(t, n.TimeoutSIGN, 2*time.Second)
 	require.Equal(t, n.TimeoutACCEPT, 2*time.Second)
@@ -22,7 +22,7 @@ func TestConfigurationDefault(t *testing.T) {
 
 //	TestConfigurationSetAndGet tests setting timeout fields and checking.
 func TestConfigurationSetAndGet(t *testing.T) {
-	n := NewNodeRunnerConfiguration()
+	n := NewIsaacConfiguration()
 	n.TimeoutINIT = 3 * time.Second
 	n.TimeoutSIGN = 1 * time.Second
 	n.TimeoutACCEPT = 1 * time.Second
