+++ lib/isaac_state.go
@@ -1,20 +1,21 @@
 // When a node receive ballot, if the ballotState is former then nodeRunnerState,
 // the ballot is ignored.
-// The node decides it is former or not by NodeRunnerState.
+// The node decides it is former or not by IsaacState.
 
 package sebak
 
 import (
 	"boscoin.io/sebak/lib/common"
+	"boscoin.io/sebak/lib/round"
 )
 
-type NodeRunnerState struct {
-	round       Round
+type IsaacState struct {
+	round       round.Round
 	ballotState sebakcommon.BallotState
 }
 
-func NewNodeRunnerState(round Round, ballotState sebakcommon.BallotState) NodeRunnerState {
-	p := NodeRunnerState{
+func NewIsaacState(round round.Round, ballotState sebakcommon.BallotState) IsaacState {
+	p := IsaacState{
 		round:       round,
 		ballotState: ballotState,
 	}
