+++ lib/isaac_state_transit_from_timeout_test.go
@@ -1,4 +1,4 @@
-// We can test that a node broadcast propose ballot or B(`EXP`) in NodeRunnerStateManager.
+// We can test that a node broadcast propose ballot or B(`EXP`) in IsaacStateManager.
 // when the timeout is expired,
 package sebak
 
@@ -13,7 +13,7 @@ import (
 
 // 1. All 3 Nodes.
 // 2. Proposer itself.
-// 3. When `NodeRunnerStateManager` starts, the node proposes ballot to validators.
+// 3. When `IsaacStateManager` starts, the node proposes ballot to validators.
 func TestStateINITProposer(t *testing.T) {
 	nodeRunners := createTestNodeRunner(3)
 
@@ -25,7 +25,7 @@ func TestStateINITProposer(t *testing.T) {
 
 	nr.Consensus().SetLatestConsensusedBlock(genesisBlock)
 
-	conf := NewNodeRunnerConfiguration()
+	conf := NewIsaacConfiguration()
 	conf.TimeoutINIT = time.Hour
 	conf.TimeoutSIGN = time.Hour
 	conf.TimeoutACCEPT = time.Hour
@@ -47,7 +47,7 @@ func TestStateINITProposer(t *testing.T) {
 
 // 1. All 3 Nodes.
 // 2. Not proposer itself.
-// 3. When `NodeRunnerStateManager` starts, the node waits a ballot by proposer.
+// 3. When `IsaacStateManager` starts, the node waits a ballot by proposer.
 // 4. But TimeoutINIT is an hour, so it doesn't broadcast anything.
 func TestStateINITNotProposer(t *testing.T) {
 	nodeRunners := createTestNodeRunner(3)
@@ -60,7 +60,7 @@ func TestStateINITNotProposer(t *testing.T) {
 
 	nr.Consensus().SetLatestConsensusedBlock(genesisBlock)
 
-	conf := NewNodeRunnerConfiguration()
+	conf := NewIsaacConfiguration()
 	conf.TimeoutINIT = time.Hour
 	conf.TimeoutSIGN = time.Hour
 	conf.TimeoutACCEPT = time.Hour
@@ -76,7 +76,7 @@ func TestStateINITNotProposer(t *testing.T) {
 
 // 1. All 3 Nodes.
 // 2. Not proposer itself.
-// 3. When `NodeRunnerStateManager` starts, the node waits a ballot by proposer.
+// 3. When `IsaacStateManager` starts, the node waits a ballot by proposer.
 // 4. But TimeoutINIT is a millisecond.
 // 5. After 200 milliseconds, the node broadcasts B(`SIGN`, `EXP`)
 func TestStateINITTimeoutNotProposer(t *testing.T) {
@@ -93,7 +93,7 @@ func TestStateINITTimeoutNotProposer(t *testing.T) {
 
 	nr.Consensus().SetLatestConsensusedBlock(genesisBlock)
 
-	conf := NewNodeRunnerConfiguration()
+	conf := NewIsaacConfiguration()
 	conf.TimeoutINIT = 1 * time.Millisecond
 	conf.TimeoutSIGN = time.Hour
 	conf.TimeoutACCEPT = time.Hour
@@ -127,8 +127,8 @@ func TestStateINITTimeoutNotProposer(t *testing.T) {
 
 // 1. All 3 Nodes.
 // 2. Proposer itself.
-// 3. When `NodeRunnerStateManager` starts, the node proposes B(`INIT`, `YES`) to validators.
-// 4. Then NodeRunnerState will be changed to `SIGN`.
+// 3. When `IsaacStateManager` starts, the node proposes B(`INIT`, `YES`) to validators.
+// 4. Then IsaacState will be changed to `SIGN`.
 // 4. But TimeoutSIGN is a millisecond.
 // 5. After 200 milliseconds, the node broadcasts B(`ACCEPT`, `EXP`)
 func TestStateSIGNTimeoutProposer(t *testing.T) {
@@ -145,7 +145,7 @@ func TestStateSIGNTimeoutProposer(t *testing.T) {
 
 	nr.Consensus().SetLatestConsensusedBlock(genesisBlock)
 
-	conf := NewNodeRunnerConfiguration()
+	conf := NewIsaacConfiguration()
 	conf.TimeoutINIT = time.Hour
 	conf.TimeoutSIGN = time.Millisecond
 	conf.TimeoutACCEPT = time.Hour
@@ -183,10 +183,10 @@ func TestStateSIGNTimeoutProposer(t *testing.T) {
 
 // 1. All 3 Nodes.
 // 2. Not proposer itself.
-// 3. When `NodeRunnerStateManager` starts, the node waits a ballot by proposer.
+// 3. When `IsaacStateManager` starts, the node waits a ballot by proposer.
 // 4. TimeoutINIT is a millisecond.
 // 5. After milliseconds, the node broadcasts B(`SIGN`, `EXP`).
-// 6. NodeRunnerState is changed to `SIGN`.
+// 6. IsaacState is changed to `SIGN`.
 // 7. TimeoutSIGN is a millisecond.
 // 8. After milliseconds, the node broadcasts B(`ACCEPT`, `EXP`).
 func TestStateSIGNTimeoutNotProposer(t *testing.T) {
@@ -203,7 +203,7 @@ func TestStateSIGNTimeoutNotProposer(t *testing.T) {
 
 	nr.Consensus().SetLatestConsensusedBlock(genesisBlock)
 
-	conf := NewNodeRunnerConfiguration()
+	conf := NewIsaacConfiguration()
 	conf.TimeoutINIT = time.Millisecond
 	conf.TimeoutSIGN = time.Millisecond
 	conf.TimeoutACCEPT = time.Hour
@@ -239,10 +239,10 @@ func TestStateSIGNTimeoutNotProposer(t *testing.T) {
 
 // 1. All 3 Nodes.
 // 2. Proposer itself at round 0.
-// 3. When `NodeRunnerStateManager` starts, the node proposes a ballot.
-// 6. NodeRunnerState is changed to `SIGN`.
-// 7. TimeoutSIGN is a millisecond.
-// 8. After milliseconds, the node broadcasts B(`ACCEPT`, `EXP`).
+// 3. When `IsaacStateManager` starts, the node proposes a ballot.
+// 4. IsaacState is changed to `SIGN`.
+// 5. TimeoutSIGN is a millisecond.
+// 6. After milliseconds, the node broadcasts B(`ACCEPT`, `EXP`).
 func TestStateACCEPTTimeoutProposerThenNotProposer(t *testing.T) {
 	nodeRunners := createTestNodeRunner(3)
 
@@ -260,7 +260,7 @@ func TestStateACCEPTTimeoutProposerThenNotProposer(t *testing.T) {
 
 	nr.Consensus().SetLatestConsensusedBlock(genesisBlock)
 
-	conf := NewNodeRunnerConfiguration()
+	conf := NewIsaacConfiguration()
 	conf.TimeoutINIT = time.Hour
 	conf.TimeoutSIGN = time.Millisecond
 	conf.TimeoutACCEPT = time.Millisecond
