+++ lib/round/round.go
@@ -1,4 +1,4 @@
-package sebak
+package round
 
 import (
 	"github.com/btcsuite/btcutil/base58"
